requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
requests>=2.32
python-multipart>=0.0.9
pdfplumber>=0.10.2
python-pptx>=0.6.23
openai>=1.25.0
tiktoken>=0.6.0
pytesseract>=0.3.11
Pillow>=10.3.0
faiss-cpu>=1.8.0
python-docx>=1.1.0
